@@ -1,7 +1,6 @@
 import 'dart:async';
 
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart' show required;
 import 'package:flutter/services.dart' show MethodChannel;
 
 class _OnAppLifecycleResumeObserver extends WidgetsBindingObserver {
@@ -32,12 +31,12 @@ class FlutterWebAuth {
   /// [callbackUrlScheme] should be a string specifying the scheme of the url that the page will redirect to upon successful authentication.
   /// [preferEphemeralSession] iOS only - Prevents the web view from using shared cookie storage.
   static Future<String> authenticate(
-      {@required String url,
-      @required String callbackUrlScheme,
+      {required String url,
+      required String callbackUrlScheme,
       bool preferEphemeralSession = false}) async {
-    WidgetsBinding.instance.removeObserver(
+    WidgetsBinding.instance!.removeObserver(
         _resumedObserver); // safety measure so we never add this observer twice
-    WidgetsBinding.instance.addObserver(_resumedObserver);
+    WidgetsBinding.instance!.addObserver(_resumedObserver);
     return await _channel.invokeMethod('authenticate', <String, dynamic>{
       'url': url,
       'callbackUrlScheme': callbackUrlScheme,
@@ -50,6 +49,6 @@ class FlutterWebAuth {
   /// terminate all `authenticate` calls with an error.
   static Future<void> _cleanUpDanglingCalls() async {
     await _channel.invokeMethod('cleanUpDanglingCalls');
-    WidgetsBinding.instance.removeObserver(_resumedObserver);
+    WidgetsBinding.instance!.removeObserver(_resumedObserver);
   }
 }
